requirements.txt
spotipy==2.23.0
pandas==2.1.4
flask==3.0.2
python-dotenv==1.0.1
pyinstaller==6.3.0 